requirements.txt
django==4.2.6
Pillow==10.1.0
djangorestframework==3.14.0
django-filter==23.3
django-star-ratings==0.9.2